use Fset and Fsymbol_value instead of ->value, changed all rep_Var implemented variables to single-arg functions (except for buffer-list which becomes (buffer-list) and (set-buffer-list NEW), and (buffer-undo-list) and (set-buffer-undo-list NEW))

+++ src/edit.h
@@ -343,7 +343,7 @@ typedef struct _VW
        C code at all; access is via the `buffer-list' variable.  */
     repv	    vw_BufferList;
 
-    short	    vw_XStepRatio, vw_YStepRatio;
+    int 	    vw_XStepRatio, vw_YStepRatio;
     short	    vw_XStep, vw_YStep;
 } VW;
